--- src/Me/MentorshipReq/index.js
@@ -1,3 +1,0 @@
-import MentorshipReq from './MentorshipReq';
-export { default as ReqContent } from './ReqContent';
-export default MentorshipReq;
