+++ src/Me/MentorshipRequests/index.js
@@ -1,0 +1,3 @@
+import MentorshipRequests from './MentorshipRequests';
+export { default as ReqContent } from './ReqContent';
+export default MentorshipRequests;
