+++ src/Me/Modals/MentorshipRequestModals/index.ts
@@ -1,3 +1,4 @@
 export { default as DeclineModal } from './DeclineModal';
 export { default as MentorshipRequest } from './MentorshipRequest';
 export { default as AcceptModal } from './AcceptModal';
+export { default as CancelModal } from './CancelModal';
